[MOV] ui_stateful: move ui_stateful plugins inside the correct folder
Task: 0
