+++ src/plugins/ui_stateful/figure.ts
