+++ src/plugins/ui_stateful/header_visibility_ui.ts
