+++ src/plugins/ui_stateful/table_computed_style.ts
