+++ src/types/getters.ts
@@ -10,26 +10,26 @@ import { FingerprintPlugin } from "../plugins/ui_core_views/fingerprint";
 import { FormulaTrackerPlugin } from "../plugins/ui_core_views/formula_tracker";
 import { HeaderSizeUIPlugin } from "../plugins/ui_core_views/header_sizes_ui";
 import { PivotUIPlugin } from "../plugins/ui_core_views/pivot_ui";
-import { CellComputedStylePlugin } from "../plugins/ui_feature/cell_computed_style";
 import { CollaborativePlugin } from "../plugins/ui_feature/collaborative";
 import { ColorThemeUIPlugin } from "../plugins/ui_feature/color_theme";
 import { DynamicTranslate } from "../plugins/ui_feature/dynamic_translate";
-import { FigureUIPlugin } from "../plugins/ui_feature/figure";
 import { GeoFeaturePlugin } from "../plugins/ui_feature/geo_features";
-import { HeaderVisibilityUIPlugin } from "../plugins/ui_feature/header_visibility_ui";
 import { HistoryPlugin } from "../plugins/ui_feature/local_history";
-import { LockSheetPlugin } from "../plugins/ui_feature/lock_sheet";
 import { PivotPresencePlugin } from "../plugins/ui_feature/pivot_presence_plugin";
 import { SortPlugin } from "../plugins/ui_feature/sort";
 import { SubtotalEvaluationPlugin } from "../plugins/ui_feature/subtotal_evaluation";
-import { TableComputedStylePlugin } from "../plugins/ui_feature/table_computed_style";
 import { UIOptionsPlugin } from "../plugins/ui_feature/ui_options";
 import { SheetUIPlugin } from "../plugins/ui_feature/ui_sheet";
 import { CarouselUIPlugin } from "../plugins/ui_stateful/carousel_ui";
+import { CellComputedStylePlugin } from "../plugins/ui_stateful/cell_computed_style";
 import { ClipboardPlugin } from "../plugins/ui_stateful/clipboard";
+import { FigureUIPlugin } from "../plugins/ui_stateful/figure";
 import { FilterEvaluationPlugin } from "../plugins/ui_stateful/filter_evaluation";
 import { HeaderPositionsUIPlugin } from "../plugins/ui_stateful/header_positions";
+import { HeaderVisibilityUIPlugin } from "../plugins/ui_stateful/header_visibility_ui";
+import { LockSheetPlugin } from "../plugins/ui_stateful/lock_sheet";
 import { GridSelectionPlugin } from "../plugins/ui_stateful/selection";
+import { TableComputedStylePlugin } from "../plugins/ui_stateful/table_computed_style";
 import { CoreGetters, PluginGetters } from "./core_getters";
 // -----------------------------------------------------------------------------
 // Getters
